refactor: Performed some cleanup& organized templates& namings for a more harmonic look and feel.
