+++ src/Controller/Data/DataController.php
@@ -6,10 +6,10 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
+#[Route('/data', name: 'app_data_index')]
 final class DataController extends AbstractController
 {
-    #[Route('/data', name: 'app_data_index')]
-    public function index(): Response
+    public function __invoke(): Response
     {
         return $this->render('data/index.html.twig', []);
     }
