+++ src/Controller/Data/DispatchAreas/ListDispatchAreasController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller\Data\Area;
+namespace App\Controller\Data\DispatchAreas;
 
 use App\DataTransferObjects\AreaListQueryParametersDTO;
 use App\Repository\DispatchAreaRepository;
@@ -9,20 +9,20 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\MapQueryString;
 use Symfony\Component\Routing\Attribute\Route;
 
-final class ListAreaController extends AbstractController
+#[Route('/data/area', name: 'app_data_dispatch_area_list')]
+final class ListDispatchAreasController extends AbstractController
 {
     public function __construct(
         private readonly DispatchAreaRepository $dispatchAreaRepository,
     ) {
     }
 
-    #[Route('/data/area', name: 'app_data_area_list')]
-    public function index(
+    public function __invoke(
         #[MapQueryString] AreaListQueryParametersDTO $query,
     ): Response {
         $paginator = $this->dispatchAreaRepository->getAreaListPaginator($query);
 
-        return $this->render('data/area/list.html.twig', [
+        return $this->render('data/dispatch_areas/list.html.twig', [
             'paginator' => $paginator,
             'search' => $query->search,
             'sortBy' => $query->sortBy,
