+++ src/Controller/DefaultController.php
@@ -6,10 +6,10 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
+#[Route('/', name: 'app_default')]
 final class DefaultController extends AbstractController
 {
-    #[Route('/', name: 'app_default')]
-    public function index(): Response
+    public function __invoke(): Response
     {
         return $this->render('default/index.html.twig');
     }
