+++ templates/_includes/card_pagination_footer.html.twig
@@ -3,5 +3,5 @@
 <div class="card-footer d-flex align-items-center">
     {{ pagination.results(paginator.numResults, paginator.pageSize) }}
 
-    {{ pagination.navbar(paginator, 'app_data_area_list') }}
+    {{ pagination.navbar(paginator, 'app_data_dispatch_area_list') }}
 </div>
