+++ templates/_includes/subnav_data.html.twig
@@ -5,13 +5,23 @@
         <div class="navbar">
             <div class="container-xl">
                 <ul class="navbar-nav">
-                    <li class="nav-item {{ route starts with 'app_data_area' ? 'active' : '' }}">
-                        <a class="nav-link" href="{{ path('app_data_area_list') }}">
-                            <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler-icons.io/i/home -->
+                    <li class="nav-item {{ route starts with 'app_data_dispatch_area' ? 'active' : '' }}">
+                        <a class="nav-link" href="{{ path('app_data_dispatch_area_list') }}">
+                            <span class="nav-link-icon d-md-none d-lg-inline-block">
                                 <twig:ux:icon name="tabler:map" width="1,5em" height="1,5em" />
                             </span>
                             <span class="nav-link-title">
-                                {{ 'link.areas'|trans }}
+                                {{ 'link.dispatch_areas'|trans }}
+                            </span>
+                        </a>
+                    </li>
+                    <li class="nav-item {{ route starts with 'app_data_hospital' ? 'active' : '' }}">
+                        <a class="nav-link" href="{{ path('app_data_hospital_list') }}">
+                            <span class="nav-link-icon d-md-none d-lg-inline-block">
+                                <twig:ux:icon name="tabler:building-hospital" width="1,5em" height="1,5em" />
+                            </span>
+                            <span class="nav-link-title">
+                                {{ 'link.hospitals'|trans }}
                             </span>
                         </a>
                     </li>
